backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
sqlalchemy==2.0.36
psycopg[binary]==3.2.3
alembic==1.14.0
pydantic-settings==2.6.1
email-validator==2.2.0
python-multipart==0.0.20
openai==1.54.4
httpx==0.27.2
pypdfium2==4.30.1
Pillow==11.0.0
